requirements.txt
Flask==3.0.3
requests==2.32.3
python-dotenv==1.0.1
SQLAlchemy==2.0.34
flask_sqlalchemy==3.1.1
flask_migrate==4.0.7
pymysql==1.1.1
boto3==1.35.20
pytest==8.3.2
flake8==7.1.1

